requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.3
gast==0.2.2
grpcio==1.26.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mock==3.0.5
numpy==1.16.2
oauthlib==3.1.0
#opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==6.2.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.13.0
#tensorboard==1.14.0
#tensorflow==1.14.0
#tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
